src/echo/service/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
cachetools==3.1.0         # via google-auth
certifi==2018.11.29       # via requests
chardet==3.0.4            # via requests
google-api-core[grpc]==1.7.0  # via google-cloud-core, google-cloud-trace, opencensus
google-auth==1.6.2        # via google-api-core
google-cloud-core==0.29.1  # via google-cloud-trace
google-cloud-trace==0.20.2  # via opencensus
googleapis-common-protos==1.5.8  # via google-api-core
grpcio-health-checking==1.13.0
grpcio==1.16.1
idna==2.8                 # via requests
opencensus[stackdriver]==0.2.0
protobuf==3.6.1           # via google-api-core, googleapis-common-protos, grpcio-health-checking
pyasn1-modules==0.2.4     # via google-auth
pyasn1==0.4.5             # via pyasn1-modules, rsa
python-json-logger==0.1.10
pytz==2018.9              # via google-api-core
requests==2.21.0          # via google-api-core
rsa==4.0                  # via google-auth
six==1.12.0               # via google-api-core, google-auth, grpcio, protobuf
urllib3==1.24.1           # via requests
